Modernize and improve the repository
Include banding for the feature.
Ensure that the analyzers can actual run on Windows.
Use OSGi for signature verification instead of jarsigner.
Use the artifact repository to determine which artifacts to test so that
the PGP signature checker works in general.
Include the m2e launcher for building locally.
Remove unused natures.
Simplify the target platform, which can be generated from the setup.